Fix #142418
