+++ src/vs/platform/workspace/common/workspace.ts
@@ -12,7 +12,16 @@ import { extname as resourceExtname, basenameOrAuthority, joinPath, extUriBiased
 import { URI, UriComponents } from 'vs/base/common/uri';
 import { createDecorator } from 'vs/platform/instantiation/common/instantiation';
 import { IEnvironmentService } from 'vs/platform/environment/common/environment';
-import { IStoredWorkspaceFolder } from 'vs/platform/workspaces/common/workspaces';
+
+export interface IRawFileWorkspaceFolder {
+	path: string;
+	name?: string;
+}
+
+export interface IRawUriWorkspaceFolder {
+	uri: string;
+	name?: string;
+}
 
 export const IWorkspaceContextService = createDecorator<IWorkspaceContextService>('contextService');
 
@@ -371,7 +380,7 @@ export class WorkspaceFolder implements IWorkspaceFolder {
 
 	constructor(
 		data: IWorkspaceFolderData,
-		readonly stored?: IStoredWorkspaceFolder
+		readonly raw?: IRawFileWorkspaceFolder | IRawUriWorkspaceFolder
 	) {
 		this.uri = data.uri;
 		this.index = data.index;
